Move CatIndex and BookIndex to fixtures dir
