+++ core/src/test/scala/pink/cozydev/protosearch/fixtures/CatIndex.scala
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package pink.cozydev.protosearch
+package pink.cozydev.protosearch.fixtures
 
 import pink.cozydev.protosearch.analysis.TokenStream.tokenizeSpaceL
+import pink.cozydev.protosearch.Index
 
 object CatIndex {
   val docs: List[List[String]] =
